+++ ostap/fitting/fit1d.py
@@ -493,7 +493,7 @@ class Shape1D_pdf(PDF1) :
             self.histo = shape
             shape      = Ostap.Math.Histo1D ( shape )
 
-        ##  iniialize the base 
+        ##  initialize the base 
         PDF1.__init__ ( self , name , xvar ) 
         
         self.__shape = shape
@@ -501,10 +501,11 @@ class Shape1D_pdf(PDF1) :
         if isinstance ( self.shape , Ostap.Math.Histo1D ) :
         
             ## create the actual pdf
-            self.pdf = Ostap.Models.Histo1D ( self.roo_name ( 'histo1_' ) , 
-                                              "Histo-1D %s" % self.name   ,
-                                              self.xvar                   ,
-                                              self.shape                  )
+            self.pdf = Ostap.Models.Histo1D (
+                self.roo_name ( 'histo1_' ) , 
+                "Histo-1D %s" % self.name   ,
+                self.xvar                   ,
+                self.shape                  )
             
         else :