+++ ReleaseNotes/release_notes.md
@@ -3,8 +3,10 @@
  1. add new test `test_fitting_fitresult.py` for variosu expressions and their uncertainties 
  1. improve `Ostap::Math::Integrator` allowig to specify the absolute and rleative precisions 
  1. move `hash_combine` from `local_hash.h` to `Ostap/Utils.h` 
- 1. improve `Ostap::Math::Piecewise` 
-   
+ 1. improve `Ostap::Math::Piecewise`          
+ 1. add helper scale factor for the Breitt-Wigner function 
+ 1. replace `ROOT.Math.erfc` with `math.erfc` in `smear`-function 
+    
 ## Backward incompatible:  
 
 ## Bug fixes: